requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
google-generativeai==0.3.2
ai4bharat-transliteration==1.1.3
numpy<2.0.0